telemetry-llm/requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
numpy==2.2.6
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
fpdf2>=2.7.8
crewai>=0.10.0
google-generativeai>=0.3.2
matplotlib>=3.8.0
python-dotenv>=1.0.0
langchain-google-genai>=0.0.7
langchain>=0.1.0
